chore(cleanup): refactor import paths to use shared module and add utility functions
