+++ src/gui/pages/account/accountpage.py
@@ -1,10 +1,10 @@
 import tkinter as tk
 from tkinter import ttk
-from utils.data.date_utils import get_iso_date
+from shared.date_utils import get_iso_date
 from gui.app.basetoplevelwindow import BaseToplevelWindow
 from gui.app.basewindow import BaseWindow
 import utils.data.database.account_utils as db_account_utils
-import utils.data.value_utils as value_utils
+import shared.value_utils as value_utils
 
 # FIXME: does not work
