+++ src/gui/pages/home/homepage.py
@@ -4,7 +4,7 @@ import locale
 import logging
 from utils.logging.logging_tools import log_fn
 from gui.app.basewindow import BaseWindow
-from utils.data.date_utils import get_month_literal
+from shared.date_utils import get_month_literal
 from utils.data.database.account_utils import get_account_data, get_total_cash
 
 logger = logging.getLogger(__name__)
