+++ src/utils/data/database/account_utils.py
@@ -7,7 +7,7 @@ import logging
 from gui.app.basewindow import BaseWindow
 from gui.pages.account.name_input_page import NameInputDialog
 from utils.data.database_connection import DatabaseConnection
-from utils.data.date_utils import get_iso_date
+from shared.date_utils import get_iso_date
 import config
